+++ src/DataPoint.cpp
@@ -2,7 +2,7 @@
 #include <cmath>
 using namespace std;
 
-DataPoint::DataPoint(string& label, vector<double>& data) {
+DataPoint::DataPoint(string& label, vector<float>& data) {
     this->label = label;
     this->data = data;
 }
@@ -10,11 +10,11 @@ DataPoint::DataPoint(string& label, vector<double>& data) {
 string DataPoint::getLabel() const {return label;}
 void DataPoint::setLabel(string label) {this->label = label;}
 
-const vector<double>& DataPoint::getData() const {return data;}
-void DataPoint::setData(int i, double val) {data[i] = val;}
+const vector<float>& DataPoint::getData() const {return data;}
+void DataPoint::setData(int i, float val) {data[i] = val;}
 
 Distance DataPoint::calcDistance(const DataPoint &x) const {
-    double d = 0;
+    float d = 0;
     for (size_t i = 0; i < data.size(); ++i) {d += abs(data[i] - x.data[i]);}
     return Distance(d, x.label);
 }
